+++ src/windows/PeripheralBase.cpp
@@ -19,6 +19,8 @@ PeripheralBase::PeripheralBase(advertising_data_t advertising_data) {
     identifier_ = advertising_data.identifier;
     address_ = advertising_data.mac_address;
     device_ = BluetoothLEDevice::FromBluetoothAddressAsync(_str_to_mac_address(advertising_data.mac_address)).get();
+    manufacturer_data_ = advertising_data.manufacturer_data;
+    connectable_ = advertising_data.connectable;
 }
 
 PeripheralBase::~PeripheralBase() {}
@@ -69,13 +71,7 @@ bool PeripheralBase::is_connected() {
     return device_ != nullptr && device_.ConnectionStatus() == BluetoothConnectionStatus::Connected;
 }
 
-bool PeripheralBase::is_connectable() {
-    // TODO its a stub, replace with real implementation
-    if (is_connected()) {
-        return false;
-    }
-    return true;
-}
+bool PeripheralBase::is_connectable() { return connectable_; }
 
 std::vector<BluetoothService> PeripheralBase::services() {
     std::vector<BluetoothService> list_of_services;
@@ -89,6 +85,8 @@ std::vector<BluetoothService> PeripheralBase::services() {
     }
     return list_of_services;
 }
+
+std::map<uint16_t, ByteArray> PeripheralBase::manufacturer_data() { return manufacturer_data_; }
 
 ByteArray PeripheralBase::read(BluetoothUUID service, BluetoothUUID characteristic) {
     GattCharacteristic gatt_characteristic = _fetch_characteristic(service, characteristic);